Adds support for the "update" API
Elasticsearch has an API for small document updates: it takes a
parameter "script" and a parameter "params" in the body and executes the
script on the document, using the parameters. Afterwards, it reindexes
the document.

This has an issue: "params" collided with the name that Eson uses for
its collections hash. This fixes this collision by renaming Esons
"params" to "parameters".

Sadly, this means that the change is breaking and has to be included in
a new release version.